requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
shapely==2.0.6
